% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookGenericCollection-methods.R
\name{as.data.frame.FacebookGenericCollection}
\alias{as.data.frame.FacebookGenericCollection}
\title{TODO: better check for null}
\usage{
as.data.frame.FacebookGenericCollection(x, row.names = FALSE,
  optional = FALSE, ...)
}
\description{
TODO: better check for null
}

