% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookPagesCollection-methods.R, R/FacebookPostsCollection-methods.R, R/generic-methods.R
\docType{methods}
\name{as.data.frame,FacebookPagesCollection,logical,logical-method}
\alias{as.data.frame}
\alias{as.data.frame,FacebookPagesCollection,logical,logical-method}
\alias{as.data.frame,FacebookPostsCollection,logical,logical-method}
\title{Returns a data frame representation of various kinds of Facebook collections}
\usage{
\S4method{as.data.frame}{FacebookPagesCollection,logical,logical}(x,
  row.names = FALSE, optional = FALSE, ...)

\S4method{as.data.frame}{FacebookPostsCollection,logical,logical}(x,
  row.names = FALSE, optional = FALSE, ...)

as.data.frame(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{A valid collection of Facebook elements}

\item{row.names}{If set to \code{TRUE}, names the rows of the returned data frame with IDs of the elements}

\item{optional}{Not used in this context.}
}
\description{
Returns a data frame representation of various kinds of Facebook collections
}
\section{Methods (by class)}{
\itemize{
\item \code{x = FacebookPagesCollection,row.names = logical,optional = logical}: to convert from a Collection of Pages.

\item \code{x = FacebookPostsCollection,row.names = logical,optional = logical}: to convert from a Collection of Pages.
}}

