% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/facebook.advanced-package.R
\docType{package}
\name{facebook.advanced}
\alias{facebook.advanced}
\alias{facebook.advanced-package}
\title{Access to Facebook API v2 via R}
\description{
This package provides a series of functions that allow R users
to access Facebook's API to get information about users and posts,
and collect public status updates that mention specific keywords.
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}

Pablo Barbera \email{pablo.barbera@nyu.edu}
}
\seealso{
\code{\link{fbOAuth}}, \code{\link{facebook.users}},
\code{\link{getPosts}}, \code{\link{searchFacebook}},
\code{\link{updateStatus}}, \code{\link{facebook.friends}},
\code{\link{getNetwork}}, \code{\link{getPages}}
}

