% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facebook.comments.R
\name{facebook.comments}
\alias{facebook.comments}
\title{Extract comments about a set of public Facebook posts}
\usage{
facebook.comments(posts, token, n = 500,
  fields = "id,from.fields(id,name),message,created_time,like_count",
  .progress = create_progress_bar())
}
\arguments{
\item{posts}{string vector or a comma-delimited string of post IDs}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{n}{numeric, maximum number of comments to return for each post.}

\item{fields}{vector or comma-delimited string with the post-level details to get.
Subfields nesting is supported.}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints
nothing to the console.}
}
\description{
\code{facebook.comments} gathers comments data from a set of public Facebook posts.
}
\details{
Due to the network-graph nature of Facebook data model,
you can always specify fields details for each field eventually nesting \code{.fields()} clauses.

For example, if you need only \code{id} and \code{name} for the \code{from} node, this clause is valid among others:
\code{from.fields(id,name)}.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
## Getting information about Facebook's Facebook Page
	load("fb_oauth")
	fb.page <- facebook.pages("9thcirclegames", token=fb_oauth)
## Getting at most 2000 comments for the ten most recent posts
	recent.posts.comments <- facebook.comments(post=fb.page$posts[10,]$id, n=2000, token=fb_oauth)
}
}
\author{
Gabriele Baldassarre
}
\seealso{
\code{\link{facebook.likes}}, \code{\link{facebook.posts}}, \code{\link{fbOAuth}}
}

