% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/facebook.friends.R
\name{facebook.friends}
\alias{facebook.friends}
\title{Extract list of friends with their information}
\usage{
facebook.friends(users = "me", token, fields = "id,name",
  .progress = create_progress_bar())
}
\arguments{
\item{users}{vector or comma-separated string of users to get friends}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{fields}{vector or comma-separated string of fields to get}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints
nothing to the console.}
}
\description{
\code{facebook.friends} retrieves information about the user's friends.
}
\details{
After the introduction of version 2.0 of the Graph API,
only friends who are using the application that you used to generate the
token to query the API will be returned.
}
\examples{
\dontrun{
## Copy and paste token created at FB Graph API Explorer
 token <- "XXXXXX"
	my.friends <- facebook.friends(token=token)
## Since users are ordered by ID, this will return 10 oldest user accounts
	head(my.friends, n=10)
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{facebook.users}}, \code{\link{fbOAuth}}
}

