% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/facebook.get.R
\name{facebook.get}
\alias{facebook.get}
\title{Extract lists of IDs}
\usage{
facebook.get(ids, token, type = "posts", n = 500, fields = "id",
  .progress = create_progress_bar())
}
\arguments{
\item{ids}{string vector comma-delimited string with IDs of the given type to extract.}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{type}{string with the child elements to get. For example \code{posts}, \code{comments}, \code{likes}.}

\item{n}{Maximum number of elements to return for each ID.}

\item{fields}{vector or comma-separated string of fields to get}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints
nothing to the console.}
}
\description{
\code{facebook.get} returns a vector of node IDs to be used in other lookup functions.
}
\details{
This generic function conveniently returns a
vector of unique IDs of \code{type} belonging to one of \code{ids}.
All \code{ids} must share the same type.

This function, bundled with a good understanding of the Facebook Data Model
enable to get virtually any kind of node from Facebook Graph API.

If partent ID ha no children of \code{type}, an error is triggered.
}
\examples{
\dontrun{
 load("fb_oauth")
 some.posts <- facebook.get(ids="9thcirclegames", type="posts", token=fb_oauth)
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{facebook.query}}, \code{\link{facebook.posts}}, \code{\link{fbOAuth}}
}

