% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/facebook.newsfeed.R
\name{facebook.newsfeed}
\alias{facebook.newsfeed}
\title{Download recent posts from the authenticated user's newsfeed}
\usage{
facebook.newsfeed(token, n = 500,
  fields = "id,from.fields(id,name),message,created_time,type,link,name",
  .progress = create_progress_bar())
}
\arguments{
\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{n}{numeric, maximum number of comments and likes to return.}

\item{fields}{vector or comma-delimited string with the post-level details to get.
\code{n}.}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints
nothing to the console.}
}
\description{
\code{faceook.newsfeed} returns a data frame with the current user' newsfeed.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
## Capture 100 most recent posts on my newsfeed and max 100 likes/comments for each one
	load("fb_oauth")
	my.newsfeed <- facebook.newsfeed(token=fb_oauth, n=200)
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{fbOAuth}}, \code{\link{getPosts}}
}

