% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facebook.query.R
\name{facebook.query}
\alias{facebook.query}
\title{Perform a generic query}
\usage{
facebook.query(query, token, endpoint = "https://graph.facebook.com/v2.3")
}
\arguments{
\item{query}{string containing the URL fragment for the FB Graph API}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{endpoint}{string with the endpoint to the FB Graph. You can optionally end the URI with the API version, here
For example \code{https://graph.facebook.com/v2.3}}
}
\description{
\code{query} performs a generic query on Facebook and returns a list reflecting the raw json tree
given by the API.
}
\details{
Write the query as you would do using the Graph Explroer at \url{https://developers.facebook.com/tools/explorer}
without including the API version number and the trailing slashes.

Please note that this function has no automatic paging support.
}
\examples{
\dontrun{
## Copy and paste token created at FB Graph API Explorer
 token <- "XXXXXX"
	results.json <- facebook.query(query="me/posts/?limit=10", token=token)
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{fbOAuth}} \code{\link{facebook.search}}
}

