% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookCommentsCollection-class.R
\docType{class}
\name{FacebookCommentsCollection-class}
\alias{FacebookCommentsCollection-class}
\title{Class for representing a Collection of Facebook comments}
\description{
Connect to Facebook Graph API, get information from a list of public Facebook comments and build a \code{\link{FacebookCommentsCollection-class}}
instance.

Use the \code{\link{FacebookCommentsCollection}} constructor for building a instance of this class, as it provides better
sanity and validation checks.
}
\section{Slots}{

\describe{
\item{\code{id}}{A character vector of the IDs of the item included in the Collection}

\item{\code{fields}}{A character vectorwith the content of any ID in the Collection. Ie \code{id, name, created_time}...}

\item{\code{token}}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth} used to fill the Collection.}

\item{\code{parent}}{A character vector of the parent IDs of the item included in the Collection or \code{NA} if the items have not inherited a parent}

\item{\code{parent.type}}{A character string with the collection type of the parent ID, if available, or NA otherwise.}
}}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
Other Facebook Collections: \code{\link{FacebookGenericCollection-class}};
  \code{\link{FacebookLikesCollection-class}};
  \code{\link{FacebookPostsCollection-class}}
}

