% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookGenericCollection.R
\name{FacebookGenericCollection}
\alias{FacebookGenericCollection}
\title{Build a generic Collection of Facebook data}
\usage{
FacebookGenericCollection(id, token, parameters = list(),
  fields = character(0))
}
\arguments{
\item{id}{A character vector or a comma-delimited string of IDs or an existing Facebook Collection of any of the supported types}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{\link{fbOAuth}}. If \code{NULL} and \code{id} is a Collection, get that one instead. Otherwise, no query is performed
to the Facebook Graph API and an empty Collection is returned}

\item{parameters}{A list of parameters to be added to the Facebook Graph API query. For more information on the
accepted parameters, see: \url{https://developers.facebook.com/docs/graph-api/using-graph-api}}

\item{fields}{A character vector or a comma-delimited string with the list of fields to get for each ID. If no value for a given fields is found, it will
be set to \code{NULL}}
}
\value{
A collection of elements in a \code{\link{FacebookGenericCollection-class}} object on on of its subclasses.
}
\description{
Connect to Facebook Graph API, get information from a list of public Facebook items and build a \code{\link{FacebookGenericCollection-class}}
instance.
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{FacebookPostset}}, \code{\link{FacebooCommentset}}, \code{\link{fbOAuth}}
}
\keyword{internal}

