% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookPagesCollection-class.R
\docType{class}
\name{FacebookPagesCollection-class}
\alias{FacebookPagesCollection-class}
\title{A collection of valid Facebook Pages.}
\description{
A collection of valid Facebook Pages.
}
\section{Slots}{

\describe{
\item{\code{pages}}{character vector with the ids/names of the pages of the collection.}

\item{\code{fields}}{character vector with the content field of each page for the collection}

\item{\code{data}}{named list representing the raw collection}
}}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}

