% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookPageset.R
\docType{class}
\name{FacebookPageset-class}
\alias{FacebookPageset}
\alias{FacebookPageset-class}
\alias{is.Pageset}
\title{A class to represent a collection of valid Facebook Pages.}
\usage{
FacebookPageset(pages, token, parameters = list(), fields = character(0))

is.Pageset(x)
}
\description{
A class to represent a collection of valid Facebook Pages.
}
\section{Functions}{
\itemize{
\item \code{FacebookPageset}: Constructs a new \code{FacebookPageset} object.

\item \code{is.Pageset}: Checks whether \code{x} is a \code{FacebookPageSet}.
}}
\section{Slots}{

\describe{
\item{\code{pages}}{character vector with the ids/names of all the pages of the collection.}

\item{\code{fields}}{character vector with the content field of each page for the collection}

\item{\code{data}}{named list representing the raw collection}
}}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}

