% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookPageset.R, R/generic-methods.R
\docType{methods}
\name{[,FacebookPageset,ANY,ANY-method}
\alias{FacebookPageset-methods}
\alias{[,FacebookPageset,ANY,ANY-method}
\alias{as.data.frame,FacebookPageset-method}
\alias{as.data.frame.FacebookPageset}
\alias{as.list,FacebookPageset-method}
\alias{as.list.FacebookPageset}
\alias{c,FacebookPageset-method}
\title{Pageset Methods}
\usage{
\S4method{[}{FacebookPageset,ANY,ANY}(x, i, j, drop)

\S4method{c}{FacebookPageset}(x, ..., recursive = FALSE)

as.list.FacebookPageset(x, ...)

\S4method{as.list}{FacebookPageset}(x, ...)

as.data.frame.FacebookPageset(x, row.names = NULL, optional = FALSE, ...)

\S4method{as.data.frame}{FacebookPageset}(x, row.names = NULL,
  optional = FALSE, ...)
}
\description{
Pageset Methods
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}

