% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FaceBookPages.R
\docType{class}
\name{FacebookPageset-class}
\alias{FacebookPageset-class}
\title{A class to represent a collection of valid Facebook Pages.}
\description{
A class to represent a collection of valid Facebook Pages.
}
\section{Slots}{

\describe{
\item{\code{pages}}{character vector with the ids/names of all the pages of the collection.}

\item{\code{fields}}{character vector with the content field of each page for the collection}

\item{\code{data}}{named list representing the raw collection}
}}

