% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookUsersCollection-class.R
\docType{class}
\name{FacebookUsersCollection-class}
\alias{FacebookUsersCollection-class}
\title{Class for representing a Collection of Facebook users}
\description{
Connect to Facebook Graph API, get public information from a list of Facebook users  nd build a \code{FacebookUsersCollection-class}
instance.

Use the \code{\link{FacebookUsersCollection}} constructor for building a instance of this class, as it provides better
sanity and validation checks.
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}

