% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookUsersCollection.R
\name{FacebookUsersCollection}
\alias{FacebookUsersCollection}
\title{Build a Collection of Facebook Users}
\usage{
FacebookUsersCollection(id, token, parameters = list(), fields = c("id",
  "name", "first_name", "last_name", "gender", "locale",
  "picture.fields(url).type(large)"), metadata = FALSE,
  .progress = create_progress_bar())
}
\arguments{
\item{id}{A character vector or a comma-delimited string of IDs or an existing Facebook Collection of any of the supported types}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{\link{fbOAuth}}. If \code{NULL} and \code{id} is a Collection, get that one instead. Otherwise, no query is performed
to the Facebook Graph API and an empty Collection is returned}

\item{parameters}{A list of parameters to be added to the Facebook Graph API query. For more information on the
accepted parameters, see: \url{https://developers.facebook.com/docs/graph-api/using-graph-api}}

\item{fields}{A character vector or a comma-delimited string with the list of fields to get for each ID. If no value for a given fields is found, it will
be set to \code{NULL}}

\item{metadata}{If set to \code{TRUE}, the medatata from each ID is pulled with the data and the \code{type} slot is fed. Please note
that setting this to \code{TRUE} will considerably slow down the query.}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints nothing to the console.}
}
\value{
A collection of pages in a \code{\link{FacebookPagesCollection-class}} object.
}
\description{
Connect to Facebook Graph API, get public information from a list of Facebook users and build a \code{\link{FacebookUsersCollection-class}}
instance.
}
\details{
\code{FacebookUsersCollection} is the constructor for the \code{\link{FacebookUsersCollection-class}} and it returns public data about users.

After version 2.0 of the Facebook API, only id, name, and picture are available
through the API as public informations. All the remaining fields will be missing unless the Application eventually asks
for extended permissions.

Due to the network-graph nature of Facebook data model,
you can always specify fields details for each field eventually nesting \code{.fields()} clauses.

For example, if you need only \code{id} and \code{source} for the \code{covern} node, this clause is valid among others:
\code{cover.fields(id,source)}.

Be careful when buinding this kind of collection starting from a \code{\link{FacebookPostsCollection}}, \code{\link{FacebookCommentsCollection}}
or a \code{\link{FacebookLikesCollection}}.

In Facebook one can publish, comment or like acting as a user or as a page. But since users and pages have different sets of fields
and you won't know in advance if a (commenting) user is a page or not, the constructor of this collection would fail.
To avoid this, if \code{id} is an instance of one of the aforementioned collection, an intermediate \code{\link{FacebookMixedCollection}}
is built with
this function will extract both. Parameters \code{users.fields} and \code{pages.fields}
are used to drive which attributes to get.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
 load("fb_oauth")

## Getting information about 9th Circle Games' Facebook Page
 fb.pages <- FacebookPagesCollection(id = c("9thcirclegames", "NathanNeverSergioBonelliEditore"), token = fb_oauth)

## Getting the commenters of the latest 10 posts
 fb.comments <- fb.pages \%>\% FacebookPostsCollection(n = 10) \%>\% FacebookCommentsCollection(fields=c("id, "from.fields(id,name)"), n = Inf)

## Build a collection of users from who actually commented those posts
 fb.commenting.users <- fb.comments -> FacebookUsersCollection()

## Convert the collection to a data frame
 fb.commenting.df <- as.data.frame(fb.commenting.users)
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{FacebookPostset}}, \code{\link{FacebooCommentset}}, \code{\link{fbOAuth}}
}

