% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getNewsfeed.R
\name{getNewsfeed}
\alias{getNewsfeed}
\title{Download recent posts from the authenticated user's newsfeed}
\usage{
getNewsfeed(token, n = 500,
  fields = "id,from.fields(id,name),message,created_time,type,link,name")
}
\arguments{
\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{n}{numeric, maximum number of comments and likes to return.}

\item{fields}{vector or comma-delimited string with the post-level details to get.
\code{n}.}
}
\description{
\code{getNewsfeed} returns a data frame with the current user' newsfeed.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
## Capture 100 most recent posts on my newsfeed and max 100 likes/comments for each one
	load("fb_oauth")
	my_newsfeed <- getNewsfeed(token=fb_oauth, n=100)
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{fbOAuth}}, \code{\link{getPosts}}
}

