% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPages.R
\name{getPages}
\alias{getPages}
\title{Extract a list of pages and posts attributes list of posts from public Facebook pages}
\usage{
getPages(pages, token, n = 100, since = NULL, until = NULL,
  feed = FALSE,
  fields = "id,username,name,about,category,description,likes,link,talking_about_count")
}
\arguments{
\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{n}{Number of posts of page to return. Note that number can be sometimes
higher or lower, depending on status of API.}

\item{since}{A UNIX timestamp or strtotime data value that points to
the start of the time range to be searched. For more information on the
accepted values, see: \url{http://php.net/manual/en/function.strtotime.php}}

\item{until}{A UNIX timestamp or strtotime data value that points to
the end of the time range to be searched. For more information on the
accepted values, see: \url{http://php.net/manual/en/function.strtotime.php}}

\item{feed}{If \code{TRUE}, the function will also return posts on the page
that were made by others (not only the admin of the page).}

\item{fields}{vector or a comma-delimited string with the page-level metadata set to get.}

\item{A}{vector or a comma-delimited string of page IDs or page name.}
}
\description{
\code{getPages} returns a list of two data frames: \code{pages} contains all metadata
of the pages while \code{posts} has a summary of the posts of the pages.
You can then use \code{\link{getPosts}} to further dig into posts details.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
## Getting information about Facebook's Facebook Page
	load("fb_oauth")
	fb_pages <- getPage(page="facebook", token=fb_oauth)
## Getting posts on 9th Circle Games page, including posts by others users
## (not only owner of page)
 page <- getPage(page="9thcirclegames", token=fb_oauth, feed=TRUE)
## Getting posts on Humans of New York page in January 2013
 page <- getPage(page="humansofnewyork", token=fb_oauth, n=1000
		since='2013/01/01', until='2013/01/31')
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{facebook.users}}, \code{\link{getPosts}}, \code{\link{fbOAuth}}
}

