% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPosts.R
\name{getPosts}
\alias{getPosts}
\title{Extract information about a set of public Facebook posts}
\usage{
getPosts(posts, token, n = 500, n.likes = n, n.comments = n,
  fields = "id,from,message,created_time,type,link,name")
}
\arguments{
\item{posts}{A vector or a comma-delimited string of post IDs}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{n}{numeric, maximum number of comments and likes to return.}

\item{n.likes}{numeric, maximum number of likes to return. Default is \code{n}.}

\item{n.comments}{numeric, maximum number of comments to return. Default is
\code{n}.}

\item{fields}{vector or comma-delimited string with the post-level details to get.
\code{n}.}
}
\description{
\code{getPosts} retrieves information about a set of public Facebook posts, including
list of comments and likes.
}
\details{
\code{getPosts} returns a list with three data frames: \code{posts},
\code{likes}, and \code{comments}. First, \code{posts} contains information
about the posts: author, creation date, id, counts of likes, comments, and
shares, etc. Second, \code{likes} is a data frame that contains names and
Facebook IDs of all the users that liked the post. Finally, \code{comments}
is a data frame with information about the comments to the posts (author,
message, creation time, id).
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
## Getting information about Facebook's Facebook Page
	load("fb_oauth")
	fb.page <- getPages("9thcirclegames", token=fb_oauth)
## Getting information and likes/comments about 10 most recent posts
	post.details <- getPosts(post=fb.page$posts[10,]$id, n=2000, token=fb_oauth)
}
}
\author{
Gabriele Baldassarre
}
\seealso{
\code{\link{facebook.users}}, \code{\link{getPage}}, \code{\link{fbOAuth}}
}

