% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getUsers.R
\name{getUsers}
\alias{getUsers}
\title{Extract information about one or more Facebook users}
\usage{
getUsers(users, token,
  user.fields = "id,name,first_name,last_name,gender,locale,picture.fields(url).type(large)",
  page.fields = "id,name,category,picture.fields(url).type(large)")
}
\arguments{
\item{users}{A vector of user or page IDs.}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}
}
\description{
\code{getUsers} retrieves public information about one or more Facebook users.

After version 2.0 of the Facebook API, only id, name, and picture are available
through the API. All the remaining fields will be missing.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
## Getting information about the authenticated user
 load("fb_oauth")
	fb <- getUsers("me,9thcirclegames", token=fb_oauth)
	fb$username
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{getFriends}}, \code{\link{getPosts}}, \code{\link{searchFacebook}}
}

