% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookPagesCollection-methods.R
\name{is.PagesCollection}
\alias{is.PagesCollection}
\title{Checks if it's a valid collection}
\usage{
is.PagesCollection(x)
}
\description{
\code{is.Pageset} checks if \code{x} is a valid collection of Facebook Pages built using \code{\link{FacebookPagesCollection}}
}

