% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookPostsCollection-methods.R
\name{is.Postset}
\alias{is.Postset}
\title{Checks if it's a valid collection of posts}
\usage{
is.Postset(x)
}
\description{
\code{is.Postset} checks if \code{x} is a valid collection of Facebook Posts built using \code{\link{FacebookPostsCollection}}
}

