% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{length}
\alias{length}
\title{Returns the number of items of a Facebook collections}
\usage{
length(x)
}
\arguments{
\item{x}{A valid collection of Facebook elements}
}
\description{
Returns the number of items of a Facebook collections
}

