% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generic-methods.R
\name{summary}
\alias{summary}
\title{Returns a summarized rapresentation of a Facebook collection}
\usage{
summary(object, ...)
}
\arguments{
\item{x}{A valid collection of Facebook elements}
}
\description{
Returns a summarized rapresentation of a Facebook collection
}

