% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookPagesCollection-methods.R, R/FacebookPostsCollection-methods.R, R/generic-methods.R
\docType{methods}
\name{as.list,FacebookPagesCollection-method}
\alias{as.list}
\alias{as.list,FacebookPagesCollection-method}
\alias{as.list,FacebookPostsCollection-method}
\title{Returns a list representation of various kinds of Facebook collections}
\usage{
\S4method{as.list}{FacebookPagesCollection}(x, ...)

\S4method{as.list}{FacebookPostsCollection}(x, ...)

as.list(x, ...)
}
\arguments{
\item{x}{A valid collection of Facebook elements}
}
\description{
Returns a list representation of various kinds of Facebook collections
}
\section{Methods (by class)}{
\itemize{
\item \code{FacebookPagesCollection}: Convert to convert from a Collection of Pages.

\item \code{FacebookPostsCollection}: to convert from a Collection of Pages.
}}

