% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/facebook.users.likes.R
\name{facebook.users.likes}
\alias{facebook.users.likes}
\title{Extract list of likes from Facebook users}
\usage{
facebook.users.likes(users, token, n = 500,
  fields = "category,name,id,created_time",
  .progress = create_progress_bar())
}
\arguments{
\item{users}{A vector or comma-delimited string with IDs or screen names.}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{n}{Maximum number of likes to return for each user.}

\item{fields}{vector or comma-separated string of fields to get}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints
nothing to the console.}
}
\description{
\code{facebook.users.likes} retrieves information about the likes from a list of Facebook IDs and/or names.
}
\details{
This is a tiny wrapper around the generic \code{\link{facebook.get}} and
requires the use of an OAuth token with the following
permissions: user_likes, friends_likes

This function requires the use of a OAuth token with user_likes
permission granted. After the introduction of version 2.0 of the Graph API,
only likes from users who are using the application that you used to generate the
token to query the API will be returned.
}
\examples{
\dontrun{
 load("fb_oauth")
 me.likes <- facebook.users.likes(users="me", token=fb_oauth)
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{facebook.friends}}, \code{\link{fbOAuth}}
}

