% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/facebook.users.R
\name{facebook.users}
\alias{facebook.users}
\title{Extract information about one or more Facebook users}
\usage{
facebook.users(users, token,
  user.fields = "id,name,first_name,last_name,gender,locale,picture.fields(url).type(large)",
  page.fields = "id,name,category,picture.fields(url).type(large)",
  .progress = create_progress_bar())
}
\arguments{
\item{users}{A vector of user or page IDs.}

\item{token}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth}.}

\item{page.fields}{vector or comma-separated string of fields to get (for page IDs)}

\item{.progress}{progress_bar object as defined in the plyr package.
By default the \code{none} progress bar is used, which prints
nothing to the console.}

\item{users.fields}{vector or comma-separated string of fields to get (for user IDs)}
}
\description{
\code{facebok.users} retrieves public information about one or more Facebook users or pages.

After version 2.0 of the Facebook API, only id, name, and picture are available
through the API. All the remaining fields will be missing.

Since users and pages have different sets of attributes and you won't know in advance if a (commenting)
user is a page or not, this function will extract both. Parameters \code{users.fields} and \code{pages.fields}
are used to drive which attributes to get.
}
\examples{
\dontrun{
## See examples for fbOAuth to know how token was created.
## Getting information about the authenticated user
 load("fb_oauth")
	fb <- facebook.users("me,9thcirclegames", token=fb_oauth)
	fb$username
}
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
\code{\link{facebook.friends}}, \code{\link{getPosts}}, \code{\link{facebook.search}}
}

