% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookPagesCollection-class.R
\docType{class}
\name{FacebookPagesCollection-class}
\alias{FacebookPagesCollection-class}
\title{A collection of valid Facebook Pages.}
\description{
A collection of valid Facebook Pages.
}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
Other Facebook Collections: \code{\link{FacebookGenericCollection-class}};
  \code{\link{FacebookPostsCollection-class}}
}

