% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookPostsCollection-class.R
\docType{class}
\name{FacebookPostsCollection-class}
\alias{FacebookPostsCollection-class}
\title{A collection of valid Facebook Posts.}
\description{
A collection of valid Facebook Posts.
}

