% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FacebookPostsCollection-class.R
\docType{class}
\name{FacebookPostsCollection-class}
\alias{FacebookPostsCollection-class}
\title{Class for representing a Collection of Facebook posts}
\description{
Connect to Facebook Graph API, get information from a list of public Facebook posts and build a \code{\link{FacebookPostsCollection-class}}
instance.

Use the \code{\link{FacebookPostsCollection}} constructor for building a instance of this class, as it provides better
sanity and validation checks.
}
\section{Slots}{

\describe{
\item{\code{id}}{A character vector of the IDs of the item included in the Collection}

\item{\code{fields}}{A character vectorwith the content of any ID in the Collection. Ie \code{id, name, created_time}...}

\item{\code{token}}{Either a temporary access token created at
\url{https://developers.facebook.com/tools/explorer} or the OAuth token
created with \code{fbOAuth} used to fill the Collection.}

\item{\code{parent}}{A character vector of the parent IDs of the item included in the Collection or \code{NA} if the items have not inherited a parent}

\item{\code{parent.collection}}{The collection used to build this one, or NA if the collection was built from atomic IDs.}

\item{\code{type}}{A factor vector with the type of each element of the collection. It's fed only if the collection is built with \code{metadata=TRUE}}
}}
\author{
Gabriele Baldassarre \email{gabriele@gabrielebaldassarre.com}
}
\seealso{
Other Facebook Collections: \code{\link{FacebookCommentsCollection-class}};
  \code{\link{FacebookGenericCollection-class}};
  \code{\link{FacebookLikesCollection-class}}
}

